Edit according to pylint
************* Module docker-judge
__init__.py:10:0: C0301: Line too long (143/100) (line-too-long)
__init__.py:18:0: C0301: Line too long (122/100) (line-too-long)
__init__.py:20:0: C0301: Line too long (120/100) (line-too-long)
__init__.py:25:0: C0301: Line too long (138/100) (line-too-long)
__init__.py:1:0: C0103: Module name "docker-judge" doesn't conform to snake_case naming style (invalid-name)
__init__.py:1:0: C0111: Missing module docstring (missing-docstring)
__init__.py:17:10: W0621: Redefining name 'settings' from outer scope (line 36) (redefined-outer-name)
__init__.py:17:0: W0102: Dangerous default value [] as argument (dangerous-default-value)
__init__.py:17:0: C0111: Missing function docstring (missing-docstring)
__init__.py:17:0: R0913: Too many arguments (7/5) (too-many-arguments)
__init__.py:17:61: W0613: Unused argument 'timeout' (unused-argument)
__init__.py:35:4: C0103: Constant name "yaml" doesn't conform to UPPER_CASE naming style (invalid-name)
__init__.py:36:4: C0103: Constant name "settings" doesn't conform to UPPER_CASE naming style (invalid-name)

-----------------------------------
Your code has been rated at 5.67/10
